requirements.txt
fastapi
uvicorn[standard]
python-dotenv
obsws-python
pydantic-settings
